requirements.txt
psd-tools
Pillow
PySimpleGUI